+++ crates/blockifier/src/fee.rs
@@ -1,5 +1,5 @@
-pub mod actual_cost;
 pub mod eth_gas_constants;
 pub mod fee_checks;
 pub mod fee_utils;
 pub mod gas_usage;
+pub mod receipt;
