+++ crates/blockifier/src/fee/fee_checks.rs
@@ -4,8 +4,8 @@ use starknet_types_core::felt::Felt;
 use thiserror::Error;
 
 use crate::context::TransactionContext;
-use crate::fee::actual_cost::TransactionReceipt;
 use crate::fee::fee_utils::{get_balance_and_if_covers_fee, get_fee_by_gas_vector};
+use crate::fee::receipt::TransactionReceipt;
 use crate::state::state_api::StateReader;
 use crate::transaction::errors::TransactionExecutionError;
 use crate::transaction::objects::{
